Module trinket_m0::adc::ctrlb

Control B

Structs

CORRENR
Value of the field

DIFFMODER
Value of the field

FREERUNR
Value of the field

LEFTADJR
Value of the field

R
Value read from the register

W
Value to write to the register

_CORRENW
Proxy

_DIFFMODEW
Proxy

_FREERUNW
Proxy

_LEFTADJW
Proxy

_PRESCALERW
Proxy

_RESSELW
Proxy

Enums

PRESCALERR
Possible values of the field PRESCALER

PRESCALERW
Values that can be written to the field PRESCALER

RESSELR
Possible values of the field RESSEL

RESSELW
Values that can be written to the field RESSEL