Module trinket_m0::tcc0::ctrla

Control A

Structs

ALOCKR
Value of the field

CPTEN0R
Value of the field

CPTEN1R
Value of the field

CPTEN2R
Value of the field

CPTEN3R
Value of the field

ENABLER
Value of the field

R
Value read from the register

RUNSTDBYR
Value of the field

SWRSTR
Value of the field

W
Value to write to the register

_ALOCKW
Proxy

_CPTEN0W
Proxy

_CPTEN1W
Proxy

_CPTEN2W
Proxy

_CPTEN3W
Proxy

_ENABLEW
Proxy

_PRESCALERW
Proxy

_PRESCSYNCW
Proxy

_RESOLUTIONW
Proxy

_RUNSTDBYW
Proxy

_SWRSTW
Proxy

Enums

PRESCALERR
Possible values of the field PRESCALER

PRESCALERW
Values that can be written to the field PRESCALER

PRESCSYNCR
Possible values of the field PRESCSYNC

PRESCSYNCW
Values that can be written to the field PRESCSYNC

RESOLUTIONR
Possible values of the field RESOLUTION

RESOLUTIONW
Values that can be written to the field RESOLUTION